Revert "feat:将领地的默认权限改为可配置#110"
This reverts commit bb2c9a687a463e7a963c62dcc3a3bfc7ffc14ca5.
