+++ src/pland/infra/Config.h
@@ -3,7 +3,6 @@
 #include "pland/Global.h"
 #include "pland/aabb/LandAABB.h"
 #include "pland/economy/EconomySystem.h"
-#include "pland/land/LandPerm.h"
 #include <unordered_set>
 #include <vector>
 
@@ -18,7 +17,7 @@ struct ForbiddenRange {
 };
 
 struct Config {
-    int              version{21};
+    int              version{20};
     ll::io::LogLevel logLevel{ll::io::LogLevel::Info};
 
     EconomyConfig economy;
@@ -71,8 +70,6 @@ struct Config {
             std::vector<ForbiddenRange>   forbiddenRanges;            // 禁止创建领地的区域
             std::map<std::string, double> dimensionPriceCoefficients; // 维度价格系数，例如维度id的1 是1.2倍 2是1.5倍
         } bought;
-
-        LandPermTable defaultperm; // 默认权限
     } land;
 
     struct {
