--- src/pland/land/LandPerm.h
@@ -1,95 +1,0 @@
-#pragma once
-
-namespace land {
-
-struct LandPermTable {
-    // 标记 [x] 为复用权限
-    bool allowFireSpread{true};           // 火焰蔓延
-    bool allowAttackDragonEgg{false};     // 点击龙蛋
-    bool allowFarmDecay{true};            // 耕地退化
-    bool allowPistonPushOnBoundary{true}; // 活塞推动
-    bool allowRedstoneUpdate{true};       // 红石更新
-    bool allowExplode{false};             // 爆炸
-    bool allowBlockFall{false};           // 方块掉落
-    bool allowDestroy{false};             // 允许破坏
-    bool allowWitherDestroy{false};       // 允许凋零破坏
-    bool allowPlace{false};               // 允许放置 [x]
-    bool allowPlayerDamage{false};        // 允许玩家受伤
-    bool allowMonsterDamage{true};        // 允许敌对生物受伤
-    bool allowPassiveDamage{false};       // 允许友好、中立生物受伤
-    bool allowSpecialDamage{false};       // 允许对特殊实体造成伤害(船、矿车、画等)
-    bool allowCustomSpecialDamage{false}; // 允许对特殊实体2造成伤害
-    bool allowOpenChest{false};           // 允许打开箱子
-    bool allowPickupItem{false};          // 允许拾取物品
-    bool allowEndermanLeaveBlock{false};  // 允许末影人放下方块
-
-    bool allowDropItem{true};          // 允许丢弃物品
-    bool allowProjectileCreate{false}; // 允许投掷物
-    bool allowRideEntity{false};       // 允许骑乘实体
-    bool allowRideTrans{false};        // 允许骑乘矿车、船
-    bool allowAxePeeled{false};        // 允许斧头去皮
-    bool allowLiquidFlow{true};        // 允许液体流动
-    bool allowSculkBlockGrowth{true};  // 允许幽匿尖啸体生长
-    bool allowMonsterSpawn{true};      // 允许怪物生成
-    bool allowAnimalSpawn{true};       // 允许动物生成
-    bool allowInteractEntity{false};   // 实体交互
-    bool allowActorDestroy{false};     // 实体破坏
-
-
-    bool useAnvil{false};             // 使用铁砧
-    bool useBarrel{false};            // 使用木桶
-    bool useBeacon{false};            // 使用信标
-    bool useBed{false};               // 使用床
-    bool useBell{false};              // 使用钟
-    bool useBlastFurnace{false};      // 使用高炉
-    bool useBrewingStand{false};      // 使用酿造台
-    bool useCampfire{false};          // 使用营火
-    bool useFlintAndSteel{false};     // 使用打火石
-    bool useCartographyTable{false};  // 使用制图台
-    bool useComposter{false};         // 使用堆肥桶
-    bool useCraftingTable{false};     // 使用工作台
-    bool useDaylightDetector{false};  // 使用阳光探测器
-    bool useDispenser{false};         // 使用发射器
-    bool useDropper{false};           // 使用投掷器
-    bool useEnchantingTable{false};   // 使用附魔台
-    bool useDoor{false};              // 使用门
-    bool useFenceGate{false};         // 使用栅栏门
-    bool useFurnace{false};           // 使用熔炉
-    bool useGrindstone{false};        // 使用砂轮
-    bool useHopper{false};            // 使用漏斗
-    bool useJukebox{false};           // 使用唱片机
-    bool useLoom{false};              // 使用织布机
-    bool useStonecutter{false};       // 使用切石机
-    bool useNoteBlock{false};         // 使用音符盒
-    bool useCrafter{false};           // 使用合成器
-    bool useChiseledBookshelf{false}; // 使用雕纹书架
-    bool useCake{false};              // 吃蛋糕
-    bool useComparator{false};        // 使用红石比较器
-    bool useRepeater{false};          // 使用红石中继器
-    bool useShulkerBox{false};        // 使用潜影盒
-    bool useSmithingTable{false};     // 使用锻造台
-    bool useSmoker{false};            // 使用烟熏炉
-    bool useTrapdoor{false};          // 使用活板门
-    bool useLectern{false};           // 使用讲台
-    bool useCauldron{false};          // 使用炼药锅
-    bool useLever{false};             // 使用拉杆
-    bool useButton{false};            // 使用按钮
-    bool useRespawnAnchor{false};     // 使用重生锚
-    bool useItemFrame{false};         // 使用物品展示框
-    bool useFishingHook{false};       // 使用钓鱼竿
-    bool useBucket{false};            // 使用桶
-    bool usePressurePlate{false};     // 使用压力板
-    bool useArmorStand{false};        // 使用盔甲架
-    bool useBoneMeal{false};          // 使用骨粉
-    bool useHoe{false};               // 使用锄头
-    bool useShovel{false};            // 使用锹
-    bool useVault{false};             // 使用试炼宝库
-    bool useBeeNest{false};           // 使用蜂巢蜂箱
-    bool placeBoat{false};            // 放置船
-    bool placeMinecart{false};        // 放置矿车
-
-    bool editFlowerPot{false}; // 编辑花盆
-    bool editSign{false};      // 编辑告示牌
-};
-
-} // namespace land
